preinstall: Move X509CertificateID types to their own file
